ml/requirements.txt
mysql-connector-python
pandas
numpy
scikit-learn
lightgbm
catboost
xgboost
python-dotenv
joblib
